+++ trunk/UldericoAlpha/Main/BoundingBox.h
@@ -9,27 +9,27 @@ namespace UldericoAlpha
 	class BoundingBox
 	{
 	private:
-		int m_width;
+		float m_width;
 
-		int m_height;
+		float m_height;
 
 	public:
 		/**
 		* Konstruktur, mit Standard Inititalisierung
 		*/
-		BoundingBox(int width = 0, int height = 0);
+		BoundingBox(float width = 0, float height = 0);
 
 		/**
 		* Getter für die Breite
 		*/
-		int GetWidth() { return m_width; };
+		float GetWidth() { return m_width; };
 
 		/**
 		* Getter für die Höhe
 		*/
-		int GetHeight() { return m_height; };
+		float GetHeight() { return m_height; };
 
-		void SetBoundingBox(int height, int width) { m_width = width; m_height = height;};
+		void SetBoundingBox(float height, float width) { m_width = width; m_height = height;};
 	};
 }
